# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ContactActivity:
    properties:
      activity_type:
        description: contact_created, note_added, email_sent, whatsapp_sent, etc.
        type: string
      contact_id:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Enterprise-grade API for managing contacts and async CSV imports.
  title: Contact Management API
  version: "1.0"
paths:
...
  /contacts/{id}/activities:
    get:
      description: Get the activity timeline and summary for a contact.
      parameters:
      - description: Contact UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Activities
      tags:
      - activities
    post:
      consumes:
      - application/json
      description: Manually log a new activity for a contact.
      parameters:
      - description: Contact UUID
        in: path
        name: id
        required: true
        type: string
      - description: Activity details (activity_type, details)
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ContactActivity'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Log Activity
      tags:
      - activities
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"